Add files via upload
In this python note book, I read the data from Kaggle, then I defined a zero-one array with a length of the input data. In this array, ones demonstrate boxes with a call of Zebra Finch and zeros implicate the boxes that do not have.

I calculated the STFT of the input data.

Then I selected the STFT elements, that were equal to one and fed them to SVM as a feature. 
SVM did not respond after 1 hour training.